don't json.dump or load for Django JSONField
