+++ src/sentry/workflow_engine/handlers/condition/group_event_handlers.py
@@ -5,7 +5,6 @@ import sentry_sdk
 from sentry.eventstore.models import GroupEvent
 from sentry.rules import MatchType, match_values
 from sentry.rules.conditions.event_attribute import attribute_registry
-from sentry.utils import json
 from sentry.utils.registry import NoRegistrationExistsError
 from sentry.workflow_engine.models.data_condition import Condition
 from sentry.workflow_engine.registry import condition_handler_registry
@@ -64,14 +63,12 @@ class EventAttributeConditionHandler(DataConditionHandler[WorkflowJob]):
 
     @staticmethod
     def evaluate_value(job: WorkflowJob, comparison: Any) -> bool:
-        comparison_dict = json.loads(comparison)
-
         event = job["event"]
-        attribute = comparison_dict.get("attribute", "")
+        attribute = comparison.get("attribute", "")
         attribute_values = EventAttributeConditionHandler.get_attribute_values(event, attribute)
 
-        match = comparison_dict.get("match")
-        desired_value = comparison_dict.get("value")
+        match = comparison.get("match")
+        desired_value = comparison.get("value")
         if not (match and desired_value) and not (match in (MatchType.IS_SET, MatchType.NOT_SET)):
             return False
