+++ src/sentry/workflow_engine/migration_helpers/issue_alert_conditions.py
@@ -6,7 +6,6 @@ from sentry.rules.conditions.every_event import EveryEventCondition
 from sentry.rules.conditions.existing_high_priority_issue import ExistingHighPriorityIssueCondition
 from sentry.rules.conditions.reappeared_event import ReappearedEventCondition
 from sentry.rules.conditions.regression_event import RegressionEventCondition
-from sentry.utils import json
 from sentry.utils.registry import Registry
 from sentry.workflow_engine.models.data_condition import Condition, DataCondition
 from sentry.workflow_engine.models.data_condition_group import DataConditionGroup
@@ -76,7 +75,7 @@ def create_event_attribute_data_condition(
     comparison = {"match": data["match"], "value": data["value"], "attribute": data["attribute"]}
     return DataCondition.objects.create(
         type=Condition.EVENT_ATTRIBUTE,
-        comparison=json.dumps(comparison),
+        comparison=comparison,
         condition_result=True,
         condition_group=dcg,
     )
